getAllFoods AJAX call up and running - returns foods accordingly
