+++ lib/index.js
@@ -1,5 +1,6 @@
-var Food = require('./food');
+var Food = require('./models/food');
 
 var someFood = new Food();
 
-var ajaxReq = require('./ajax_requests')
+const bs = require('./ajax_requests/food');
+bs();
